Offers Module ADOP
-Added new CW721-Offers Module ADOP panels
-Setup panels for: place-offer, accept-offer, cancel-offer
-Added panels to new-panel-inventory

Next Steps:
This module has a few usage restrictions which need to be declared in the structure for the ADOP. Quick notes were left for reference.

There are some implementation questions regarding the setup and usage of the offers module which need to be discussed with backend. Panel adjustments may have to be made depending on if there are any contract changes.