Panel Renaming Controls
Panel name assignment base features. To transition from the assigned temp UUID tag for declared names.

Next Steps:
Requires alterations to panel reading for assignment in the msgConstructor. (Requires adding the name field or an alteration to the $ID field.) (Use of $ID may already be assigned in current usage and needs to be confirmed, likely easier to add a $panel-name field)