Panel Selector UX Repairs
-Resolving definition loss error on panel-name changes
-Resetting persistent search filtering issues with panel selector system

Notes:
Due to the nature of duplicating the panel and then removing previous copy, renamed panels shift to becoming the latest item. This can cause a panel that has been renamed to be pushed to the bottom of the stack.
This can be resolved by: changing the reference type from object name to an meta-field identifier (like $panel-name); Adding a display sorting order, or changing to an inline injection method.
This has not been addressed here as it is not priority, but is being noted for project tracking to be addressed.